Windows: Fixed `os::mkdir()` to support long paths.
This commit fixes some long path problems on Windows by replacing the
use of the deprecated POSIX function `mkdir` with the explicit Unicode
function `CreateDirectoryW` on Windows. Paths are resolved to be be
absolute and then prepended with `\\?\`, which forces long path support
without requiring a registry or group policy setting.

Because `CreateDirectory` will fail with "permission denied" on paths
such as `C:\`, `os::exists` is used to check if the path exists before
attempting to create it.

Review: https://reviews.apache.org/r/60302/