Ported the net::mac to the Windows Platform.
Note that, on Windows, net::mac takes a MAC address rather than a
network name. See associated changes to net::links for an explanation
of why this is the case.

Just for sanity, even though we take a MAC address as input, we go
ahead and validate the input, validating that the mac address passed
in is actually valid on the current system.