Ported the net::links routine to Windows platform.
Note that, rather than returning an interface name, the Windows
version will return the MAC address. The actual interface name,
on Windows, is a GUID (since there can be multiple wireless
interfaces, for example). However, it's not possible to get the
GUID from the command line (via 'ipconfig').

As a result, the MAC address of the interface (also unique) is
used to uniquely identify a particular interface on the system.

Review: https://reviews.apache.org/r/59200