Windows: Replaced `ping` with `Start-Sleep`.
This `SLEEP_COMMAND` macro represents a command that should sleep. On
POSIX systems, a `sleep` binary is available, but on Windows, `sleep` is
a `cmd.exe` feature that requires a TTY. Previously, this was replaced
with `ping` as it approximates sleep, except that it produces noisy
output. We replace it with the `Start-Sleep` PowerShell command, which
does not require a TTY, and is always available.

Review: https://reviews.apache.org/r/60325/