Document that some FetcherTest tests won't work on Windows platform.
Tests for tar, gzip, and such won't be working on Windows for
the time being. Thoughts are to provide this capability to the
Fetcher in a cross-platform manner via a programmatic code library
rather than Linux-specific command line tools (tar, gzip, etc).

In the short term, zip and unzip will work since PowerShell can
support that natively.